go/kreq-gen: set default value of CMAKE_GO_BINARY
this allows developer to build the tree without relying on build.sh.
cmake script is supposed to be self-contained in the sense that it
should be able to detect the dependencies, and fail whe generating
the building system if any of the dependencies is not satisified.
so, in this change, a cached variable is added to store the path to
"go" compiler. this allows us to build the tree with cmake, and
to override the path by passing -DCMAKE_GO_BINARY=/path/to/go.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>