compat/test: store string in json by its value
before this change, we store a string into JSON document with storing
a view of it, namely, to store a pointer to it. but the string being
stored is local variable and is allocated on stack. after the
`if (value.IsString())` ends, the string is destroyed.
that's why we'd have "heap-use-after-free" reported by ASan when testing
using `ctest --output-on-failure -R test_compat_rpunit`.

per https://rapidjson.org/md_doc_tutorial.html#CreateString, we are
allowed to store the string by copying it.

after this change, an additional `doc` parameter is passed into
`perturb()` function, and the `doc` is used for getting the allocator
from it, so the stored string is alive as long as the `doc` is alive.
and the test passes without issues.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>