c/backend: info log boostrap offset ranges
Recently we ran into an issue in a wiped off node where we suspect
controller recovery racing with data partition recovery. This might have
resulted in executing partition opts that would otherwise have been
skipped in bootstrapping process.

To debug such racy situations, adds the begin and end offset of each ntp
as it bootstraps. It is done once per ntp boostrap if there are some
deltas to recover, so should be ok to do at info level.

Each subsequent partition op that is executed has its offset logged in
the controller backend, so we can check if it falls in the offset range.