rpk: send cluster config updates to leader
...rather than to any node, and relying on
forwarding to leader.

This gets a more deterministic result in the edge
case of no-op changes where rpk cannot determine
client-side that it is a no-op, and relies on
the admin API code to do that.

Because this check is done on in the admin API handler,
it happens on whichever node the request lands on, which
could be an out of date follower, in which case we might
write a (harmless, redundant) additional config version to
the controller log.

That can trip up an automated test.

Fixes https://github.com/redpanda-data/redpanda/issues/6163