cloud_storage: Add prefetch tracker
The prefetch_tracker class is used by the prefetch mechanism in
remote_partition. It is passed from one remote_segment_batch_reader
to another and tracks number of consumed bytes total and inside
current segment. The operator() returns true if the prefetch needs to be
invoked.

The tracker has three parameters:
- flag that enables or disables prefetching
- per-consumer size threshold - number of bytes after which the
  prefetching is enabled
- per-segment remaining bytes size limit, the prefetch is triggered when
  the consumer consumed enough bytes so the remaining bytes are blow the
limit.