tests/cluster_metadata: give rpk more oportunities to ask different node
When one of the nodes is isolated from the other it may be the case that
rpk will always contact the isolated node when requesting cluster
metadata, first node to contact is selected randomly.
To give it more chance to ask other nodes made querying more
often and increased a timeout.

Signed-off-by: Michal Maslanka <michal@redpanda.com>