tests: wait until `license` feature is fully supported
We may start logging the nag message on a node that has activated the
`license` while there are others in the cluster that have not yet
activated the feature. This patch updates the behavior so we wait for
the feature to be active on all nodes before proceeding.

Without the fix, when applying #6003, I saw 1/10 failures of this test.
With it, I ran the test with 50/50 successful runs.

Fixes #6143