tests: generalize OMB smoke test to include other services
It is a useful thing to have quick tests that validate
that the testing components themselves are working as
expected, because not all of these components are used
in PR tests -- some are only used on nightly tests.

This way we don't risk committing things that break
the services used in the nightlies without seeing
the failure in the PR CI run.