tests: fetch multiple metric families
This commit adds a method the Redpanda service that allows for fetching
samples for multiple metric families: "metrics_samples". Previously,
call sites that required multiple metric families had to make separate
calls to "metrics_sample" for each metric family required. This was
inconvenient for waiting on a set of metrics to become available.