cluster: Even distribution of cgrp partitions
Partition allocation domain is introduced. There are two kinds of
domains: one domain is `common`, and all other domains are priority.

Node allocaion space is now calculated per allocation domain.
When a partition is allocated in a priority domain,
a modified set of constraints is used that only considers allocations
made within that domain. Partition allocations in the common domain
keeps considering allocations made in all domains, priority or not.

A single built-in priority allocation domain is introduced, the one for
__consumer_offsets topic.

The existing unit tests are updated to match the previous behaviour,
all partition allocations are made in the common domain there.

(cherry picked from commit efd48975668a254d59aaa5881ea324baeaf55296)