treewide: use named mutexes
All mutexes are named mutexes.

When a mutex fails / is broken, the name will be included in the
exception message.

This is analagous to how we use named semaphores, universally, to aid in
debugging. Each mutex gets a name. The naming convention follows
ssx::semaphore's, except each name has the prefix 'mtx:'.