s/disk_log_impl: advance segment generation after adjacent compaction
When adjacent segments are compacted we must advance resulting segment
generation. It was previously observed that the segment which was
a result of adjacent segments compaction had exactly the same generation
as the segment from before compaction. This may lead to incorrect
truncation of a segment as physical offset valid for the segment from
before adjacent compaction is most likely invalid for the segment after
adjacent compaction.

Fixes: #6253

Signed-off-by: Michal Maslanka <michal@redpanda.com>