redpanda: shutdown partitions before transaction coordinator
Transaction coordinator may depend on some underlying partition
operation to finish. In order to prevent the `tx__gateway_frontend` from
waiting for partition operation to finish before shutdown we stop all
the partitions in one of the first shutdown steps. This way all the
partition operations are interrupted before `tx_gateway_frontend` is
asked for shutdown.

Fixes: #6193

Signed-off-by: Michal Maslanka <michal@redpanda.com>