tools/offline_log_viewer: added serde decoding for topic_commands
the new code tries to mimic the structure of the adl decoding functions.
the new code is aware of the version number embedded in the binary stream and will skip the field if the version is not supported, while reporting the error.

(cherry picked from commit 021317ad3315d891041ee38b24bf880ba7faf14c)