archival: Fix lifetime issue
Convert add_ntp_archiver to coroutine and hold a gate.
Also, add a check that prevents code from using null pointer in case
if ntp is no longer available. Modify 'stop' method by invoking 'stop'
methods of archivers directly and after closing the gate. When the
gate is closed it's guaranteed that reconciliation is not running in the
background fiber so the race condition is prevented.