net: introduced invalid request error
Introduced `invalid_request_error` exception allowing a protocol handler
to inform the server that connection has to be dropped because request
payload is invalid f.e. request size limits are exceeded.

The `invalid_request_error` when handled by the server will terminate
the connection but it will not be recognized as unhandled server error.

Signed-off-by: Michal Maslanka <michal@redpanda.com>