tests: rpk group describe retry in ManyPartitionsTest
This is a bit different from other failures that are
retried inside rpk.py.  There's probably an underlying
rpk bug to be fixed here in the error handling, but
this makes the test handle rpk's current behavior, which
can be a very slow timeout when the whole cluster was
just restarted.