cluster: Do not write preparing state on disk
We can reduce the number of disk operation if we will not write preparing
state on disk. But after it we should ans to client when we sure that tx
will be recommited after fail. We can garanie it only if we ans after
marking tx prepared. Becase after fail tx will be recommited again and
client will see expected bechavior.