rptest: Set timeout in transactional producer
The timeout parameter which is passed to the init_transactions is
eventually passed to rd_kafka_init_transactions. The default value that
python wrapper is using is -1. This commit allows to set the timeout
value explicitly. The default value is set to 120 seconds.

(cherry picked from commit 065f56f9ec2385f0a7a66108422cab3ad05716b2)