tests/admin_op_fuzzer: add custom shuffling of the rpk brokers list
This is a bit of an ugly hack to deal with the consequences of
https://github.com/redpanda-data/redpanda/issues/6317. If the
first broker in the list is suspended, rpk won't try other brokers
and will fail straight away. This is a problem even if we shuffle
the brokers list (as redpanda.brokers_list() does) and add retries -
there is a small chance that the first broker will be the bad one
in all retries! To mitigate this we deteriministically choose the
first broker so that it is always different from the previous
invocation and shuffle the rest.

Fixes https://github.com/redpanda-data/redpanda/issues/6075

(cherry picked from commit fd7a66c385d88882bd819c779eb2734173011eeb)