rpk: add confirmation prompt to iotune
iotune tuner checked if the io-config file was
already present before executing, which means if
the user wanted to rerun iotune (after a fix or a
mount change), it would have to delete the file
manually.

Now the user will have to confirm via a CLI prompt
if they want to proceed with an Iotune execution
and we warn them that this will delete previous
iotune results.

In the future, we should change this overwriting
approach in the redpanda-iotune binary to
something like appending each test result to a
list.