cloud_storage: Add tests that cover replaced segments
Test serialization/deserialization of the replaced segments collection.
The collection is a multi-map, so redpanda should be able to deal with
JSON objects with duplicate keys.

The duplicate keys can be added when several segments with the same base
offset are replaced. This is an unlikely scenario but it should be
tested anyway.