cluster: Add new commands to archival STM
Add 'update_start_offset_cmd' used by retention, and
'cleanup_metadata_cmd' used by the GC procedure.

Implement removal backlog method that returns a list of segments that
can be collected by the GC.

The start_offset field is added to the snapshot. The list of segments
that snapshot has is used for both actual and replaced segments.