cloud_storage: re-set reader after exception
This commit changes the exception handling of
partition_record_batch_reader_impl, wrapper for
remote_segment_batch_reader, in order to catch all exceptions and re-set
the segment's reader.

The change is necessary because the read path is not exception safe,
which means that an exception thrown mid-read can leave the parser in an
invalid state. An example of such a scenario has been observed in #6054:
a bad_alloc exception was thrown from read_iobuf_exactly at some point
after the first iteration. The input stream was advanced, but the
parser's offset was not updated. When the request was re-tried using the
same remote_segment_batch_reader it resulted in a short read as the
storage layer was not aware of the position of the input stream.