archival: adds search for next compacted segment
to eventually support uploading compacted topics to SI, we need to be
able to search for the next compacted segment after a given offset.

this change adds a function to search for the next such segment, reusing
the code from the normal segment search to be able to get file positions
inside a segment and create an upload candidate.