tests: fixed race condition by starting all consumer before producers
The race condition in a tests caused them to fail as one of the consumer
consumed all of the messages before the others joined the group. As the
test is based on assumption that the consumers will consume from roughly
the same number of partitions it failed. Fixed the race condition by
starting producer after all of the consumers are started. This way we
are certain that consumers are members before any messages appear in the
partitions.

Fixes: #5885
Fixes: #5952

Signed-off-by: Michal Maslanka <michal@redpanda.com>