admin: add swagger spec for peer_status request
This commit updates the swagger specification to introduce a peer_status
debug request. The request returns the current status from
node_status_table of the peer with the associated id. It is only to
be used for ducktape testing. Proper integration of metadata from
the node status subsystem with the public admin API will follow at some
later point.