archival: enables uploading compacted segments
the upload loop creates two contexts and uses them to track two kinds of
segment uploads independently.

an upload context enables tracking uploads of a certain kind of segment
by storing the start and end offsets as well as the kind of segment and
the current upload count, this allows the upload loop to upload
different kinds of segments while staying in their individual boundaries
as well as honoring the global concurrency limits.

the new upload type introduced here is for compacted segments, which are
re-uploaded to potentially replace already existing non-compacted
segments in the cloud. the re-upload is always aligned to the boundaries
of the segments in cloud, and can span more than one non-compacted
segment by concatenating multiple compacted segments into a single
segment.