tests: extend consumer timeout in test with kills
test_write_with_node_failures randomly kills nodes while a verifiable
consumer is running against the cluster. Hence, the consumer group
members need to react to the broker failures. If we get unlucky, and the
test continously kills the group coordinator broker, then there may be
long periods of time when no data is consumed. This non-determinism
coupled with a short consumer timeout (30s) makes the test flaky.

This commit increases the consumer timeout to 90s in order to increase
the chances of all the required data being consumed before the timeout.