storage: apply limits to segment.bytes
To avoid introducing incompatibilities with applications
that have existing segment.bytes settings during topic
creation, we do not refuse topic creations/alterations
that violate these limits.  Instead, we accept the user input,
but clamp it when applying the segment size limit, so that
the effective segment size remains within the bounds even
if the apparent segment.bytes property violates it.

Fixes: https://github.com/redpanda-data/redpanda/issues/6036