tests/services: two phase stop for kafka cli svc
This commit makes the KafkaCliConsumer stop procedure two phase.
We first try to terminate it gracefully and allow it to leave the
consumer group. If that doesn't work we SIGKILL the process to ensure
we don't end up with a stray consumer like in #6490.

(cherry picked from commit 73150027676ca2ad026e0869a15039f0c6f2820a)