serde: pass header by value to serde_async_read
Previously, it was very easy to write subtly broken
serde_async_read coroutines that tried to refer
to `hdr` after a scheduling point.  They would
occasionally get zeroed memory, which isn't an obvious
failure because zero is a valid version for most structs,
but causes logical misbehavior when the rest of the
application is expecting the fields on latest-version
structs to be present + decoded.