cloud_storage: check that we have serialized the whole manifest
It turns out that if the underlying buffer object throws (e.g. a
bad_alloc exception), std::ostream swallows the exception and sets the
"badbit". If we don't check it, this can lead to the serialized manifest
being truncated and to corrupt manifests being uploaded to the cloud
storage. To prevent that we check that std::ostream is in a good state
after serializing manifests.

More info: https://en.cppreference.com/w/cpp/io/ios_base/iostate

Conflicts:
	src/v/cloud_storage/tx_range_manifest.cc