archival: Use manifest instance from the archival_metadata_stm
Currently, archival subsystem and archival_metadata_stm are using their
own copies of the partition manifest.
- ntp_archiver_service downloads its own copy on startup, when the
  segment is uploaded it adds its metadata to archival_metadata_stm and
also to its own copy of the manifest, then it re-uploads its own copy of
the manifest back to S3.
- archival_metadata_stm stores its own copy in the snapshot and
  periodically updates it using archival metadata batches from the log.

The manifest from archival_metadata_stm is used inside the
remote_partition so it's a source of truth for shadow indexing.
The recovery and read-replicas depend on manifest in the remote bucket.

This approach has several problems.
- If the manifest in the cloud is corrupted redpanda will crash after
  leadership is esablished.
- If the manifest in the cloud is deleted redpanda will continue
  uploading segments but it will reupload the manifest which misses old
segments.
- The memory usage is doubled.

This commit removes partition_manifest instance from ntp_archiver
replacing it with the instance of the manifest that belongs to the
archival_metadata_stm. This means that we have only one copy of the
manifest which is used in all subsystems.

This also eliminates the necessity to download the manifest once the
leadership is established and ntp_archiver is created. Whith this
cahnges the ntp_archiver will use the state from archival_metadata_stm.