redpanda: make app shutdown order more explicit
I will soon need to tweak the application startup ordering to support
assigning node IDs before starting most subsystems. This reordering is
tricky to get right, especially considering the current startup/shutdown
sequences:

The RP application maintains a list of lambdas to run on shutdown, which
we currently use to destruct several Services. Also incorporated into
the list is some handling of nuanced shutdown dependency handling. For
instance, we shutdown Kafka and internal RPC inputs up front before
destructing everything else, to ensure a bounded amount of work to
complete as we shutdown.

The registered lambdas are run in reverse order upon shutting down.
This seems error prone and hard to extend. This commit makes the
shutdown sequence a bit more explicit by running the bodies of some such
lambdas in the shutdown hook directly.

The new explicit shutdown sequence in this commit focuses only on the
more nuanced shutdown ordering. It thus ignores the previous
interleaving of some Service destructions, as the destructions seem
irrelevant to the shutdown ordering requirements.

The new shutdown sequence assumes that each shutdown event can run
whether or not we've completed the startup process (previously this was
handled by only registering a shutdown event after completing certain
startup events). This commit addresses this by:
- guarding the running of the shutdown events on construction of each
  Service by checking if each service is initialized,
- making some adjustments to ensure that a shutdown that happens before
  starting up is a no-op.