storage: adds reader for multiple segments
A new segment reader handle is added which iterates over multiple
segments. It also takes a start and end filepos, the first segment in the
list is read from the start filepos and the last segment in the list is
read upto the end filepos.

The intermediate segments are read fully. The new utility does not quite
behave internally like a segment reader handle but it exposes the same
interface, to allow the SI remote utilities to upload multiple segments
using the same interface that is used for a single segment.