cluster: refactor feature table to its own library
This is a rather invasive change for such a small class,
but it should be worthwhile long term.

Much like the `config` library, the `features` library
might be accessed from anywhere else in redpanda, and needs
to be pulled out of cluster to avoid circular dependencies, for
example when something in `raft` needs to check a feature.

This commit makes no logical changes whatsoever: it is only
refactoring paths.