c/node_status: add node_status_provider subsystem
This commit introduces a new node staus subsystem that has the purpose
of providing information on the status of a node's peers outside of
Raft.

There's two components to it:
* node_status_backend: Runs on shard 0 of every node and performs the
  following:
    1. Maintains a list of peers for the current node node
    2. Send a periodic node_status RPC to all known peers
    3. Update node_status_table with the collected information from
       the responses
* node_status_table: Runs on all shards and exposes node status
  information to other subsystems.

This is a lower level abstraction compared to the existing
health_monitor subsystem. It's intended usage is within the RPC and
consesus layers.