tests/partition_balancer/full_nodes: wait for the producer to finish
Waiting for greater than 70% average disk usage is a bit flaky because
of changing distribution of fallocated chunks among segments. Instead we
simply wait for the producer to finish. If insufficient data will be
produced to have a meaningful test run, we'll fail at the next stage
where we wait for the "stalled" balancer status.