r/vote_stm: use cached term value to update leader state
`raft::vote_stm` is optimized not to hold the consensus `_op_lock` while
waiting for the vote request replies. The locking is done in two phases.
In a first phase a lock is acquired to make a snapshot of a current
state and create a `vote_request` then in the second phase candidate
state is updated according to the election result.

The two phase approach make it possible for other request to be
responded by the node after `vote_stm` released the first phase lock.
In this situation a node may no longer be a candidate and the term might
have already changed. In this case a candidate even if an election was
successful must not become a leader.

The second phase lock is kept after the successful election up the point
where a newly elected leader dispatches request to replicate
configuration. After this (when waiting for replication result) a lock
is released.

Newly introduced assertion checking `_confirmed_term` was preceded with
a condition validating if current term is equal to the term that was
cached during the second phase lock. This approach is incorrect as there
might have been term update between the phases or during configuration
replication. In other words an assertion was validating the result of an
election which was won in `current_term-1` with the assumption it was
won in `current_term`.

Additionally not using a cached term value may introduce a confusion as
log entries coming from the `vote_stm` were using current consensus term
but not the one which the election was actually won in. This way the
logs might have been suggesting that there were two active leaders in
the same term (which is not true).

Signed-off-by: Michal Maslanka <michal@redpanda.com>