Add an API and metric for available memory
Available memory is free memory (e.g., as reported by the
seastar allocator) *plus* memory which can be reclaimed.

That is, it is the amount of memory that new allocations can
reasonably expect to be able to use without suffering an allocation
failure.

This change adds an available_memory class, which exposes an API to
return the estimated amount of available memory, as
described above. It keeps a list of reclaimers which must
register with it to provide their estimate of reclaimable memory.

Currently there is only one such reclaimer: the batch cache,
which already knows how much memory it is holding.

Fixes redpanda-data/core-internal#103.