tx/tests: unit test coverage for aborted tx
As we start relying more on the list of aborted txs spanning a given
offset range, it is critical to have some regression tests so that we
do not break the semantics by mistake.

This patch adds some common scenarios for aborted transactions spanning
single and multiple segments.

This is purely a test change and makes no changes to the core tx logic.