cloud_storage/storage tests: fixes segment upload test
segment upload test has to be reduced in scope, because the imposter we
use does not yet support byte expectations. it has to be extended to
support byte expectations (it only supports strings as of now) to be
able to assert that we are actually downloading a binary segment file.

in addition fixes the storage stop to use defer instead of stopping at
the end of the test, so that if test fails we do not get extra assertion
failures due to open storage.