Ensure background reclaimer is always notified
In batch_cache::put we did not notify the reclaimer on an early exit
from the function, when a large batch was added, probably because
this code path was added later, introducing the early return.

This means that if this code path is always/usually triggered, the
background reclaimer won't start up and we may run low on free memory.
Ultimately this memory can still be reclaimed but at a latency cost
since it happens "inline" with an allocation request, likely during
request processing.

Fix it by ensuring notify() is called on all paths.

Fixes #6573.

(cherry picked from commit e3c0a77b1ef777799c36c67f353097b9806ad51f)