archival: adds segment reupload collection utility
Segment collector allows collecting log segments for re-uploads. It is
intended for use with compacted segment reuploads, while also collecting
some extra information related to segment alignment with the partition
manifest.

All compacted segments starting from last uploaded compacted offset upto
a given size are collected and concatenated into one segment to be
uploaded to cloud.

A start and end offset is also collected, this is intended to be used to
"trim" the collected segments so that the start and end align with the
segments in partition manifest.