cloud_storage: wrap directory walker cb in coroutine::lambda
This is a defensive measure.

Use of a bare lambda is currently safe because of how list_directory
is implemented: the lambda gets moved into a stream object internally,
which keeps the lambda and its captures alive as long as it will
be used.

However, we generally avoid passing reference captures into lambdas
without some mitigation, and it isn't strictly guaranteed that this
directory listing function will behave this way, so we add a
ss::coroutine::lambda wrapper to make it more obviously correct.