tests: selectively enable segment size jitter on stress tests
The purpose of disabling this by default in tests is to enable
tests with strict success conditions (like "produce until N
segments are uploaded") to proceed deterministically.

For tests which do not do that kind of inspection and are more
general stress tests, it makes more sense to have segment jitter
size turned on, to increase coverage and to be closer to
a production system.

(cherry picked from commit 9a8ad515d9e07d6238db049449638e7a72bb8620)