redpanda: fast-activate all features in redpanda_thread_fixture
...by creating a specializatino of wire_up_and_start
for use in "unit" tests that construct whole application
objects.

This simplifies test execution for any components that
rely on a particular feature flag: they do not have to
wait for the controller to come up.

Coverage of the slower real-world path where the controller
initializes features is still there in integration tests.