c/types: init local target props on upgrade
This commit changes the deserialisation logic of
cluster::topic_configuration in order to handle upgrades.

Consider the following scenario: we are upgrading from a redpanda
version which is not aware of the retention.local.target.* topic
configuration options. If the topic in question is enabled for remote
writes to cloud storage, the existing retention.* topic configs control
the local retention. In the new version, for topics with remote write
enabled, retention.local.target.* should control local retention instead
and retention.* should be used for total retention.

This change in behaviour is addressed by introducing custom
deserialisation logic. When reading a cluster::topic_configuration
object encoded by a version which predates retention.local.target* we
use the retention.bytes and retention.ms values to initialise the
retention.local.target.bytes and retention.local.target.ms. We also
leave the retention.* configs unitialised as previously there were no
retention limits on remote partitions.

The same trick is also performed for ADL encoding.