r/replicate_batcher: handle timeout in replicate batcher
Added handling of timeout in replicate batcher. The timeout may be
triggered in either one of the two replicate stages. The timeout can
happen either when waiting for item to be cached, then the first stage
is timedout with timeout exception or during the second phase, then the
first phase is resolved and the second phase result is set to timeout
error code.

When the timeout error is triggered during the first phase the item is
not cached and the buffers are freed immediately. If an exception happen
during the second phase processing then the replicate request isn't
canceled but the result future is resolved by the timeout timer.
Replicate batcher memory units are released only after the replicate
call finishes even if future was already resolved. This way in case of
replication being slow and timing out we will still be propagating
backpressure without allocating memory for pending requests.

When the timeout is triggered in the second phase of replicate request
processing the replication outcome is undetermined. It may fail or
succeed.

Signed-off-by: Michal Maslanka <michal@redpanda.com>