cloud_storage: Add start-offset to manifest
Add start_offset field to the partition manifest. The field is the same
as in archival_metadata_stm. It's updated when first segment is added
and when the method 'advance_start_offset' is called.

Backward compatibility is handled in the following way. If the
start_offset field can't be found inside the manifest it's computed from
the list of segments.

The empty manifest has star_offset set to model::offset::min(). When all
data is removed from the manifest the field gets reset. When the first
segment is added to the empty manifest the field gets set to its
base_offset.

Remote partition uses this value through the manifest maintained by the
archival_metadata_stm.

The advance_start_offset can only be called with offsets matching the
base_offsets of the segments in the cloud. The only exception is offsets
above last segment (this is needed to allow removal of all segments in
the manifest).