test: remove tiny falloc config in test_cpu_limited
This used to be needed because partition_allocator applied
a condition that at least one falloc-step of space was needed
per partition, so the test de-facto disabled this check
by setting a tiny falloc step.

However, since storage_resources was added with its dynamic
falloc size, that space check is no longer applied at partition
creation time, so it is unnecessary for the test to set
a custom falloc size.

This tiny falloc step was a suspected contributor to occasional
failures of this test with unstable raft leadership (perhaps
the tiny falloc step was making the storage layer too slow).

Related: https://github.com/redpanda-data/redpanda/issues/6545