rptest: bump initial version in upgrade test
This commit updates the initial version used in the controller upgrade
tests to 22.2.5, which is the latest patch version of the previous main
release at the thime of this commit. The change is done to bring this
test in line with our upgrade policy: "upgrades to the next major
version must be done from the latest previous major version".