archival: fix bug SI mode selection
Previously, topic level overrides of remote read/write were not honoured
in some cases. For example, if remote write was enabled at the cluster
level (cloud_storage_enable_remote_write=true) and disabled at the topic
level (redpanda.remote.write=false), remote writes would still be
performed. This is due to a bug in the logic which chooses the SI mode:
if none of the topic level overrides are true,
'get_shadow_indexing_mode` returns a std::nullopt which causes the
default value of the properties to be used. The default value is
whatever the cluster level config is set to.

This commit fixes the issue. Note that we may still return a
std::nullopt if both remote read and write are unspecified at the topic
level. This means that the topic will always follow the cluster level
remote read/write configuration.