kafka/tests: fix incorrect expected errors setting on `create_topics`
This test appears to have been written expecting 2 errors,
then had the problematic paths commented out, so now it
expects 2 errors but gets zero.

This doesn't hurt until there actually _is_ an error in the test,
at which point we don't get a proper failure report, we just
get the test framework complaining that it expected 2 errors
but only got 1.