kafka/tests: fix assertion in `list_offsets`
The test data generation creates a situation where
we may not reason robustly about the timestamp
results at all, because the query time and the
record batch time are both taken independently
from the wallclock, and the record batch times
are rewound to earlier than the wallclock when
generated.

There will be a followup to fix the test data
generation.