archival: Fix assertion in archival_policy
The 'get_file_range' method can scan segments to find the last record
batch that should be included into S3 upload. This is done using a file
stream that reads directly from disk. The file size is checked in the
begining and can change while the scan is perfomed. In this case the
file size will be outdated and this will trigger an assertion.

This is fixed by using latest file size value instead of the cached one.