cloud_storage: implement S3 data deletion on topic deletion
This only happens if remote.write is enabled (either by topic
override or by global setting), and the topic is not a read
replica topic.

There are still some edge cases where objects may be left
behind in S3: this doesn't break anything but it does consume
some space.  Because manifests and segments are updated non-atomically,
we will need a 'scrub' type operation to resolve these cases.