r/types: renamed last reply timestamp field
The
`follower_index_metadata::last_received_append_entries_reply_timestamp`
should be updated after each successful follower reply. The name of
field indicated that it is only in case of the response on
`append_entries_request`. Changed the field name to
`last_received_reply_timestamp` in order to correctly reflect
the update scenarios.

Signed-off-by: Michal Maslanka <michal@redpanda.com>